Refine shared hooks and clean lint issues
