+++ src/App.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
-import { AuthProvider, useAuth } from './contexts/AuthContext';
+import { AuthProvider } from './contexts/AuthContext';
+import { useAuth } from './hooks/useAuth';
 import { EventWarehouseProvider } from './contexts/EventWarehouseContext';
 import Login from './components/Login';
 import Register from './components/Register';
