+++ src/components/Login.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/useAuth';
 import { LogIn, Eye, EyeOff } from 'lucide-react';
 
 interface LoginProps {
