+++ src/components/PhotoCapture.tsx
@@ -1,105 +1,12 @@
 import { useEffect, useRef, useState } from 'react';
 import { Camera, ImageOff, Loader2, RefreshCcw } from 'lucide-react';
-
-export type Roi = { xPct: number; yPct: number; wPct: number; hPct: number };
-
-export const ROI_BARCODE: Roi = { xPct: 0.1, yPct: 0.55, wPct: 0.8, hPct: 0.25 };
-export const ROI_TEXT_TOP: Roi = { xPct: 0.08, yPct: 0.1, wPct: 0.84, hPct: 0.28 };
-export const ROI_LOT: Roi = { xPct: 0.08, yPct: 0.85, wPct: 0.84, hPct: 0.12 };
-
-export interface RoiCropResult {
-  barcode?: Blob;
-  textTop?: Blob;
-  lot?: Blob;
-  hints: {
-    roi: {
-      barcode: Roi;
-      textTop: Roi;
-      lot: Roi;
-    };
-    expected: {
-      barcodeSymbologies: string[];
-      keywords: string[];
-    };
-  };
-}
+import { getRoiCrops } from './photoCaptureUtils';
+import { type RoiCropResult } from './photoCaptureTypes';
 
 interface PhotoCaptureProps {
   file: File | null;
   onChange: (file: File | null, crops: RoiCropResult | null) => void;
   disabled?: boolean;
-}
-
-export async function getRoiCrops(file: File): Promise<RoiCropResult | null> {
-  if (typeof window === 'undefined') return null;
-
-  const hints: RoiCropResult['hints'] = {
-    roi: {
-      barcode: ROI_BARCODE,
-      textTop: ROI_TEXT_TOP,
-      lot: ROI_LOT
-    },
-    expected: {
-      barcodeSymbologies: ['ITF-14', 'EAN-13'],
-      keywords: ['RAINDANCE', 'NAMAQUA', 'FILLING DATE']
-    }
-  };
-
-  async function cropImage(image: HTMLImageElement, roi: Roi): Promise<Blob | undefined> {
-    try {
-      const canvas = document.createElement('canvas');
-      const sx = Math.floor(image.width * roi.xPct);
-      const sy = Math.floor(image.height * roi.yPct);
-      const sw = Math.floor(image.width * roi.wPct);
-      const sh = Math.floor(image.height * roi.hPct);
-      canvas.width = Math.max(1, sw);
-      canvas.height = Math.max(1, sh);
-      const ctx = canvas.getContext('2d');
-      if (!ctx) return undefined;
-      ctx.drawImage(image, sx, sy, sw, sh, 0, 0, canvas.width, canvas.height);
-      const blob = await new Promise<Blob | null>((resolve) => {
-        canvas.toBlob((value) => resolve(value), 'image/jpeg', 0.7);
-      });
-      return blob ?? undefined;
-    } catch (error) {
-      console.warn('ROI crop failed', error);
-      return undefined;
-    }
-  }
-
-  const image = await loadImage(file);
-  const [barcode, textTop, lot] = await Promise.all([
-    cropImage(image, ROI_BARCODE),
-    cropImage(image, ROI_TEXT_TOP),
-    cropImage(image, ROI_LOT)
-  ]);
-
-  if (!barcode && !textTop && !lot) {
-    return { hints };
-  }
-
-  return {
-    barcode,
-    textTop,
-    lot,
-    hints
-  };
-}
-
-async function loadImage(file: File): Promise<HTMLImageElement> {
-  return new Promise((resolve, reject) => {
-    const url = URL.createObjectURL(file);
-    const image = new Image();
-    image.onload = () => {
-      URL.revokeObjectURL(url);
-      resolve(image);
-    };
-    image.onerror = (error) => {
-      URL.revokeObjectURL(url);
-      reject(error);
-    };
-    image.src = url;
-  });
 }
 
 export default function PhotoCapture({ file, onChange, disabled }: PhotoCaptureProps) {
