+++ src/components/Register.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/useAuth';
 import { UserPlus, Eye, EyeOff } from 'lucide-react';
 
 interface RegisterProps {
