+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
-import { QueryClient, QueryClientProvider } from './lib/queryClient.tsx';
+import { QueryClient, QueryClientProvider } from './lib/queryClient';
 
 const queryClient = new QueryClient();
