+++ supabase/functions/submit-count/index.ts
@@ -336,7 +336,7 @@ Deno.serve(async (req: Request) => {
 
     try {
       await supabase.rpc('refresh_counts_totals_mv');
-    } catch (_refreshError) {
+    } catch {
       // Ignore refresh failures to keep submission fast
     }
